feat: contactBody component
