+++ packages/apps/crm/src/components/organisms/index.js
@@ -14,3 +14,4 @@ export {default as ClientHeader} from './ClientHeader/ClientHeader';
 export {default as ClientBody} from './ClientBody/ClientBody';
 export {default as ClientBottom} from './ClientBottom/ClientBottom';
 export {default as ContactHeader} from './ContactHeader/ContactHeader';
+export {default as ContactBody} from './ContactBody/ContactBody';
