+++ packages/apps/crm/src/screens/contact/ContactDetailsScreen.js
@@ -1,27 +1,12 @@
-import React, {useEffect, useMemo} from 'react';
+import React, {useEffect} from 'react';
 import {ScrollView, StyleSheet, View} from 'react-native';
-import {
-  Screen,
-  HeaderContainer,
-  Text,
-  LabelText,
-  NotesCard,
-  DropdownCardSwitch,
-  CircleButton,
-} from '@axelor/aos-mobile-ui';
+import {Screen, HeaderContainer, CircleButton} from '@axelor/aos-mobile-ui';
 import {
   useSelector,
   HeaderOptionsMenu,
-  useTranslator,
   useDispatch,
 } from '@axelor/aos-mobile-core';
-import {
-  ContactHeader,
-  DropdownContactView,
-  DropdownEventView,
-  PartnerCard,
-} from '../../components';
-import {getLastEvent, getNextEvent} from '../../utils/dateEvent';
+import {ContactBody, ContactHeader} from '../../components';
 import {fetchContactEventById} from '../../features/eventSlice';
 import {fetchPartner} from '../../features/partnerSlice';
 import {getContact} from '../../features/contactSlice';
@@ -29,20 +14,8 @@ import {getContact} from '../../features/contactSlice';
 const ContactDetailsScreen = ({navigation, route}) => {
   const idContact = route.params.idContact;
   const dispatch = useDispatch();
-  const I18n = useTranslator();
   const {mobileSettings} = useSelector(state => state.config);
-  const {listEventContact} = useSelector(state => state.event);
-  const {partner} = useSelector(state => state.partner);
   const {contact} = useSelector(state => state.contact);
-
-  const lastEventClient = useMemo(() => {
-    return getLastEvent(listEventContact);
-  }, [listEventContact]);
-
-  const nextEventClient = useMemo(() => {
-    return getNextEvent(listEventContact);
-  }, [listEventContact]);
-
   React.useLayoutEffect(() => {
     navigation.setOptions({
       headerRight: () => (
@@ -77,97 +50,7 @@ const ContactDetailsScreen = ({navigation, route}) => {
         fixedItems={<ContactHeader />}
       />
       <ScrollView>
-        {partner?.isCustomer === true && partner?.isProspect === false && (
-          <PartnerCard
-            style={styles.item}
-            partnerFullName={partner.simpleFullName}
-            partnerReference={partner.partnerSeq}
-            partnerAdress={partner.mainAddress?.fullName}
-            partnerFixedPhone={partner.fixedPhone}
-            partnerEmail={partner.emailAddress?.address}
-            partnerPicture={partner.picture}
-            onPress={() =>
-              navigation.navigate('ClientDetailsScreen', {
-                idClient: partner.id,
-              })
-            }
-          />
-        )}
-        {partner?.isCustomer === false && partner?.isProspect === true && (
-          <PartnerCard
-            style={styles.item}
-            partnerFullName={partner.simpleFullName}
-            partnerReference={partner.partnerSeq}
-            partnerScoring={partner.leadScoringSelect || 0}
-            partnerAdress={partner.mainAddress?.fullName}
-            partnerFixedPhone={partner.fixedPhone}
-            partnerEmail={partner.emailAddress?.address}
-            partnerPicture={partner.picture}
-            onPress={() =>
-              navigation.navigate('ProspectDetailsScreen', {
-                idProspect: partner.id,
-              })
-            }
-          />
-        )}
-        <NotesCard title={I18n.t('Crm_Notes')} data={contact.description} />
-        <View style={styles.container}>
-          <DropdownCardSwitch
-            styleTitle={styles.textTitle}
-            dropdownItems={[
-              {
-                title: I18n.t('Crm_Contact'),
-                key: 5,
-                childrenComp: (
-                  <DropdownContactView
-                    address={contact.mainAddress?.fullName}
-                    fixedPhone={contact.fixedPhone}
-                    mobilePhone={contact.mobilePhone}
-                    emailAddress={contact.emailAddress?.address}
-                    webSite={contact.webSite}
-                  />
-                ),
-              },
-              {
-                title: I18n.t('Crm_GeneralInformation'),
-                key: 6,
-                childrenComp: (
-                  <View>
-                    {contact.user?.fullName && (
-                      <LabelText
-                        title={I18n.t('Crm_AssignedTo')}
-                        iconName={'user-tie'}
-                        value={contact.user?.fullName}
-                      />
-                    )}
-                    {contact.language?.name && (
-                      <LabelText
-                        title={I18n.t('Crm_Language')}
-                        value={contact.language?.name}
-                      />
-                    )}
-                    {!contact.user?.fullName && !contact.language?.name && (
-                      <View>
-                        <Text>{I18n.t('Crm_NoGeneralInformation')}</Text>
-                      </View>
-                    )}
-                  </View>
-                ),
-              },
-              {
-                title: I18n.t('Crm_Events'),
-                key: 7,
-                childrenComp: (
-                  <DropdownEventView
-                    lastEvent={lastEventClient}
-                    nextEvent={nextEventClient}
-                    navigation={navigation}
-                  />
-                ),
-              },
-            ]}
-          />
-        </View>
+        <ContactBody navigation={navigation} />
       </ScrollView>
       <View style={styles.bottomContainer}>
         <CircleButton
